+++ src/database/db_helpers/index.js
@@ -1,16 +1,18 @@
 // export the database functions
 module.exports = {
+
     // - GET users
-    checkUser: require('./checkUser').checkUser,
-    checkPassword: require('./checkUser').checkPassword,
+    checkUser: require('./checkUser'),
+    checkPassword: require('./checkPassword'),
     getUserId: require('./getUserId'),
     // - POST users
     addUser: require('./addUser'),
     
     // - GET talks
     checkTalk: require('./checkTalk'),
-    getAllTalks: require('./getTalks').getAllTalks,
-    upComingTalks: require('./getTalks').upComingTalks,
+    getAllTalks: require('./getAllTalks'),
+    upcomingTalks: require('./upcomingTalks'),
     // - POST talks
     addTalkToDatabase: require('./addTalkToDatabase'),
+
 };
