+++ src/views/helpers/uppercase.js
