+++ tests/index.js
@@ -1,6 +1,8 @@
 // run all tests at the same time
 module.exports = {
-    functions: require("./functions.test"),
+    controllers: require("./controllers.test"),
+    handlebars: require("./handlebars.test"),
+    dom: require("./dom.test"),
     routes: require("./routes.test"),
     database: require("./database.test")
 };
